backend/requirements.txt
Django==4.2
djangorestframework
django-cors-headers
gunicorn
requests
boto3
psycopg2-binary
celery